+++ src/components/navigation/Footer.stories.tsx
@@ -70,7 +70,6 @@ export const LogoOnly: Story = {
 
 export const CopyrightOnly: Story = {
   args: {
-    logo: null,
     copyright: "Company",
   },
 };