+++ changelog.md
@@ -6,10 +6,11 @@ SciReactUI Changelog
 --------------------
 
 ### Added
--
+- Logo component
 
 ### Fixed
 - Styles added to Navbar and Footer incorrectly remove built in styles.
+- Logo not appearing when no dark src set in dark mode.
 
 ### Changed
 - Breadcrumbs component takes optional linkComponent prop for page routing.