+++ changelog.md
@@ -2,6 +2,18 @@ SciReactUI Changelog
 ====================
 
 
+[v0.2.1alpha] - 2025-??-??
+--------------------
+
+### Added
+- Logo component, to easily add the theme logo to anywhere
+
+### Fixed
+- 
+
+### Changed
+- ImageColorSchemeSwitch takes parameter to interchange images on theme mode.
+
 [v0.2.0] - 2025-06-11
 ---------------------